feat(strategy): full-chain strategy lifecycle management
- Add DB migration 0013: strategy status, description, tags, symbols fields
- Add StrategyStatus enum with transitions (Draft→Backtesting→Deployed→Running→Paused→Archived)
- Add StrategySummaryRow with to_domain() for type-safe DB→domain conversion
- Refactor strategy_repository: extract from service, add find_all/find_by_id/insert/update/delete
- Extend StrategyService: get_strategies, save_strategy (upsert), lifecycle actions
- Fix get_strategies: use row.to_domain() instead of deserializing params JSON as StrategyParams
- Fix id type: i64→i32 to match PostgreSQL SERIAL (INT4)
- Extend Tauri commands: lifecycle (deploy/start/stop/pause/resume/archive)
- Add StrategyStatusTag, StrategyDetailPanel, PerformanceChart components
- Refactor Strategy.vue to use Pinia store, remove inline data fetching
- Add strategy.ts service module, extend types.ts with StrategyParams status fields
- Fix StrategyParams TS type: add status/description/tags/symbols
- Clippy clean: redundant closures, type-complexity, is_some_and, unnecessary casts
- 295 tests pass, 0 failures

+++ crates/data-puller/src/lib.rs
@@ -190,7 +190,7 @@ impl DataPuller {
         bar: String,
         config: CandlePullConfig,
     ) {
-        let interval = Duration::from_secs(config.interval_secs as u64);
+        let interval = Duration::from_secs(config.interval_secs);
         info!("Starting candle pull: {} / {} every {:?}", symbol, bar, interval);
 
         loop {
@@ -247,7 +247,7 @@ impl DataPuller {
         symbol: String,
         config: TickerPullConfig,
     ) {
-        let interval = Duration::from_secs(config.interval_secs as u64);
+        let interval = Duration::from_secs(config.interval_secs);
         info!("Starting ticker pull: {} every {:?}", symbol, interval);
 
         loop {
@@ -302,7 +302,7 @@ impl DataPuller {
         symbol: String,
         config: IntervalConfig,
     ) {
-        let interval = Duration::from_secs(config.interval_secs as u64);
+        let interval = Duration::from_secs(config.interval_secs);
         info!("Starting funding rate pull: {} every {:?}", symbol, interval);
 
         loop {
@@ -345,7 +345,7 @@ impl DataPuller {
         symbol: String,
         config: IntervalConfig,
     ) {
-        let interval = Duration::from_secs(config.interval_secs as u64);
+        let interval = Duration::from_secs(config.interval_secs);
         info!("Starting mark price pull: {} every {:?}", symbol, interval);
 
         loop {
@@ -386,7 +386,7 @@ impl DataPuller {
         repo: Arc<MarketDataRepository>,
         config: IntervalConfig,
     ) {
-        let interval = Duration::from_secs(config.interval_secs as u64);
+        let interval = Duration::from_secs(config.interval_secs);
         info!("Starting account balance pull every {:?}", interval);
 
         loop {
@@ -429,7 +429,7 @@ impl DataPuller {
         repo: Arc<MarketDataRepository>,
         config: IntervalConfig,
     ) {
-        let interval = Duration::from_secs(config.interval_secs as u64);
+        let interval = Duration::from_secs(config.interval_secs);
         info!("Starting position pull every {:?}", interval);
 
         loop {
